test split + block unhealthy validator group
